fixed problems from last and refactored to use decimal instead of doubble

+++ Models/Product.cs
@@ -6,11 +6,11 @@ namespace ErpCli.Models
         public string? ItemNumber { get; set; }
         public string? Name { get; set; }
         public string? Description { get; set; }
-        public double? Price { get; set; }
-        public double? Cost { get; set; }
-        public double? ProfitPercent => (Price.HasValue && Cost.HasValue && Cost.Value != 0) ? ((Price.Value - Cost.Value) / Cost.Value) * 100 : (double?)null;
+        public decimal? Price { get; set; }
+        public decimal? Cost { get; set; }
+        public decimal? ProfitPercent => (Price.HasValue && Cost.HasValue && Cost.Value != 0) ? ((Price.Value - Cost.Value) / Cost.Value) * 100 : (decimal?)null;
         public string? Location { get; set; }
-        public double StockQuantity { get; set; }
+        public decimal StockQuantity { get; set; }
         public Unit Unit { get; set; }
     }
     public enum Unit
